backend/requirements.txt
alembic==1.4.0
astroid==2.3.3
autopep8==1.5
Babel==2.8.0
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
